+++ swafe/runner.py
@@ -3,23 +3,8 @@ from botocore.vendored.requests.exceptions import ReadTimeout
 from botocore.exceptions import ClientError
 from task import ActivityTask, DecisionTask
 from lib import swf
-import uuid
 from daemon import Daemon
 from multiprocessing import Process
-
-
-def start_workflow(domain, workflowType, taskList, activity_input):
-    try:
-        response = swf.start_workflow_execution(
-            domain=domain,
-            workflowId='ingestion-'+str(uuid.uuid4()),
-            workflowType=workflowType,
-            taskList=taskList,
-            input=activity_input
-        )
-        print "Workflow requested: ", response
-    except ClientError as e:
-        print "Workflow excecution already started: ", e.response.get("Error", {}).get("Code")
 
 
 class Runner(Daemon):
